move staff to accessor.hpp
