+++ include/gridtools/stencil_composition/stencil_composition.hpp
@@ -23,8 +23,6 @@
 #include "accessor.hpp"
 #include "caches/define_caches.hpp"
 #include "computation.hpp"
-#include "define_accessors.hpp"
-#include "esf.hpp"
 #include "global_parameter.hpp"
 #include "grid.hpp"
 #include "make_computation.hpp"
