+++ unit_tests/stencil_composition/structured_grids/test_domain_reassign_expparam.cpp
@@ -17,7 +17,6 @@
 #include <gridtools/stencil_composition/accessor.hpp>
 #include <gridtools/stencil_composition/arg.hpp>
 #include <gridtools/stencil_composition/computation.hpp>
-#include <gridtools/stencil_composition/define_accessors.hpp>
 #include <gridtools/stencil_composition/expandable_parameters/make_computation.hpp>
 #include <gridtools/stencil_composition/grid.hpp>
 #include <gridtools/stencil_composition/make_computation.hpp>
