+++ unit_tests/stencil_composition/structured_grids/test_local_domain.cpp
@@ -26,7 +26,6 @@
 #include <gridtools/stencil_composition/accessor.hpp>
 #include <gridtools/stencil_composition/arg.hpp>
 #include <gridtools/stencil_composition/axis.hpp>
-#include <gridtools/stencil_composition/define_accessors.hpp>
 #include <gridtools/stencil_composition/grid.hpp>
 #include <gridtools/stencil_composition/intermediate.hpp>
 #include <gridtools/stencil_composition/make_stage.hpp>
